+++ gate-web/src/main/groovy/com/netflix/spinnaker/gate/security/WebSecurityAugmentor.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.internal.security
+package com.netflix.spinnaker.gate.security
 
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder
