+++ gate-web/src/main/groovy/com/netflix/spinnaker/gate/security/oauth2/OAuth2SecurityConfig.groovy
@@ -14,23 +14,18 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.internal.security.oauth2
+package com.netflix.spinnaker.gate.security.oauth2
 
-import com.netflix.spinnaker.gate.security.oauth2.IdentityResourceServerTokenServices
-import com.netflix.spinnaker.internal.security.WebSecurityAugmentor
-import com.netflix.spinnaker.internal.security.x509.X509AuthenticationProvider
+import com.netflix.spinnaker.gate.security.WebSecurityAugmentor
 import com.netflix.spinnaker.security.AuthenticatedRequest
 import com.netflix.spinnaker.security.User
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
-import org.apache.http.impl.client.BasicAuthCache
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.DependsOn
-import org.springframework.core.annotation.Order
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.authentication.AuthenticationProvider
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
@@ -38,12 +33,9 @@ import org.springframework.security.config.annotation.authentication.builders.Au
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.core.Authentication
 import org.springframework.security.core.userdetails.UserDetailsService
-import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer
-import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfiguration
 import org.springframework.security.oauth2.provider.authentication.OAuth2AuthenticationProcessingFilter
 import org.springframework.security.oauth2.provider.token.DefaultAccessTokenConverter
 import org.springframework.security.oauth2.provider.token.UserAuthenticationConverter
-import org.springframework.security.web.authentication.preauth.x509.X509AuthenticationFilter
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter
 import org.springframework.web.client.RestOperations
 import org.springframework.web.client.RestTemplate
