+++ gate-web/src/main/groovy/com/netflix/spinnaker/gate/security/saml/SAMLSecurityController.groovy
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.internal.security.saml
+package com.netflix.spinnaker.gate.security.saml
 
 import com.netflix.spinnaker.gate.config.Headers
 import com.netflix.spinnaker.gate.services.internal.KatoService
-import com.netflix.spinnaker.internal.security.AnonymousAccountsService
-import com.netflix.spinnaker.internal.security.anonymous.AnonymousSecurityConfig
+import com.netflix.spinnaker.gate.security.AnonymousAccountsService
+import com.netflix.spinnaker.gate.security.anonymous.AnonymousSecurityConfig
 import com.netflix.spinnaker.security.User
 import groovy.util.logging.Slf4j
 import org.opensaml.saml2.binding.encoding.HTTPRedirectDeflateEncoder
