+++ gate-web/src/main/groovy/com/netflix/spinnaker/gate/security/x509/X509SecurityConfig.groovy
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.internal.security.x509
+package com.netflix.spinnaker.gate.security.x509
 
-import com.netflix.spinnaker.internal.security.AnonymousAccountsService
-import com.netflix.spinnaker.internal.security.WebSecurityAugmentor
+import com.netflix.spinnaker.gate.security.AnonymousAccountsService
+import com.netflix.spinnaker.gate.security.WebSecurityAugmentor
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression
 import org.springframework.context.annotation.Configuration
